Use evn variable for secret and fix empty state param
